requirements.txt
requests==2.32.3
requests-cache==1.2.1
tenacity==9.0.0
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.2.2
